Changed the coordsToString() method
Changed this method so that it will return a string and also a list of bad variables that have not yet been assigned. This is good because then we can create synthetic MetaPost path variable asignment statements to create these bad varaibles. MetaPost does not like when a variable not declared when used.